split lib to multiple files
