+++ default_apps/manual_pipeline/manual_pipeline_app.py
@@ -1,167 +1,4 @@
 @dataclass
-class SyftLink(Jsonable):
-    @classmethod
-    def from_file(cls, path: str) -> SyftLink:
-        if not os.path.exists(path):
-            raise Exception(f"{path} does not exist")
-        with open(path, "r") as f:
-            return cls.from_url(f.read())
-
-    def from_path(path: str) -> SyftLink | None:
-        parts = []
-        collect = False
-        for part in str(path).split("/"):
-            # quick hack find the first email and thats the datasite
-            if collect:
-                parts.append(part)
-            elif validate_email(part):
-                collect = True
-                parts.append(part)
-
-        if len(parts):
-            sync_path = "/".join(parts)
-            return SyftLink.from_url(f"syft://{sync_path}")
-        return None
-
-    @classmethod
-    def from_url(cls, url: str | SyftLink) -> SyftLink:
-        if isinstance(url, SyftLink):
-            return url
-        try:
-            # urlparse doesnt handle no protocol properly
-            if "://" not in url:
-                url = "http://" + url
-            parts = urlparse(url)
-            host_or_ip_parts = parts.netloc.split(":")
-            # netloc is host:port
-            port = 80
-            if len(host_or_ip_parts) > 1:
-                port = int(host_or_ip_parts[1])
-            host_or_ip = host_or_ip_parts[0]
-            if parts.scheme == "https":
-                port = 443
-
-            return SyftLink(
-                host_or_ip=host_or_ip,
-                path=parts.path,
-                port=port,
-                protocol=parts.scheme,
-                query=getattr(parts, "query", ""),
-            )
-        except Exception as e:
-            raise e
-
-    def to_file(self, path: str) -> bool:
-        with open(path, "w") as f:
-            f.write(str(self))
-
-    def __init__(
-        self,
-        protocol: str = "http",
-        host_or_ip: str = "localhost",
-        port: int | None = 5001,
-        path: str = "",
-        query: str = "",
-    ) -> None:
-        # in case a preferred port is listed but its not clear if an alternative
-        # port was included in the supplied host_or_ip:port combo passed in earlier
-        match_port = re.search(":[0-9]{1,5}", host_or_ip)
-        if match_port:
-            sub_server_url: SyftLink = SyftLink.from_url(host_or_ip)
-            host_or_ip = str(sub_server_url.host_or_ip)  # type: ignore
-            port = int(sub_server_url.port)  # type: ignore
-            protocol = str(sub_server_url.protocol)  # type: ignore
-            path = str(sub_server_url.path)  # type: ignore
-
-        prtcl_pattrn = "://"
-        if prtcl_pattrn in host_or_ip:
-            protocol = host_or_ip[: host_or_ip.find(prtcl_pattrn)]
-            start_index = host_or_ip.find(prtcl_pattrn) + len(prtcl_pattrn)
-            host_or_ip = host_or_ip[start_index:]
-
-        self.host_or_ip = host_or_ip
-        self.path: str = path
-        self.port = port
-        self.protocol = protocol
-        self.query = query
-
-    def with_path(self, path: str) -> Self:
-        dupe = copy.copy(self)
-        dupe.path = path
-        return dupe
-
-    @property
-    def query_string(self) -> str:
-        query_string = ""
-        if len(self.query) > 0:
-            query_string = f"?{self.query}"
-        return query_string
-
-    @property
-    def url(self) -> str:
-        return f"{self.base_url}{self.path}{self.query_string}"
-
-    @property
-    def url_no_port(self) -> str:
-        return f"{self.base_url_no_port}{self.path}{self.query_string}"
-
-    @property
-    def base_url(self) -> str:
-        return f"{self.protocol}://{self.host_or_ip}:{self.port}"
-
-    @property
-    def base_url_no_port(self) -> str:
-        return f"{self.protocol}://{self.host_or_ip}"
-
-    @property
-    def url_no_protocol(self) -> str:
-        return f"{self.host_or_ip}:{self.port}{self.path}"
-
-    @property
-    def url_path(self) -> str:
-        return f"{self.path}{self.query_string}"
-
-    def to_tls(self) -> Self:
-        if self.protocol == "https":
-            return self
-
-        # TODO: only ignore ssl in dev mode
-        r = requests.get(  # nosec
-            self.base_url, verify=verify_tls()
-        )  # ignore ssl cert if its fake
-        new_base_url = r.url
-        if new_base_url.endswith("/"):
-            new_base_url = new_base_url[0:-1]
-        return self.__class__.from_url(
-            url=f"{new_base_url}{self.path}{self.query_string}"
-        )
-
-    def __repr__(self) -> str:
-        return f"<{type(self).__name__} {self.url}>"
-
-    def __str__(self) -> str:
-        return self.url
-
-    def __hash__(self) -> int:
-        return hash(self.__str__())
-
-    def __copy__(self) -> Self:
-        return self.__class__.from_url(self.url)
-
-    def set_port(self, port: int) -> Self:
-        self.port = port
-        return self
-
-    @property
-    def sync_path(self) -> str:
-        return self.host_or_ip + self.path
-
-    @property
-    def datasite(self) -> str:
-        return extract_leftmost_email(str(self))
-
-
-@dataclass
 class SyftVault(Jsonable):
     mapping: dict
 
@@ -268,41 +105,6 @@ def create_datasite_import_path(datasite: str) -> str:
     import_path = datasite.replace("@", ".at.")
     # Append '.datasets' to the import path
     return import_path
-
-
-@dataclass
-class DatasiteManifest(Jsonable):
-    datasite: str
-    file_path: str
-    datasets: dict = field(default_factory=dict)
-    code: dict = field(default_factory=dict)
-
-    @classmethod
-    def load_from_datasite(cls, path: str) -> DatasiteManifest | None:
-        datasite_path = Path(os.path.abspath(path))
-        manifest_path = datasite_path / "public" / "manifest" / "manifest.json"
-        try:
-            manifest = DatasiteManifest.load(manifest_path)
-            return manifest
-        except Exception:
-            pass
-        return None
-
-    class Dataset:
-        sync_path: str
-
-
-def extract_leftmost_email(text: str) -> str:
-    # Define a regex pattern to match an email address
-    email_regex = r"[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+"
-
-    # Search for all matches of the email pattern in the text
-    matches = re.findall(email_regex, text)
-
-    # Return the first match, which is the left-most email
-    if matches:
-        return matches[0]
-    return None
 
 
 def attrs_for_datasite_dataset_import(module, sync_import_path: str) -> dict[str, Any]:
@@ -463,11 +265,6 @@ def to_safe_function_name(name: str) -> str:
     if name and name[0].isdigit():
         name = "_" + name
     return name
-
-
-def markdown_to_html(markdown_text):
-    html = markdown.markdown(markdown_text)
-    return html
 
 
 @dataclass
@@ -673,274 +470,6 @@ class CodeResults:
 
         df = pd.DataFrame(table_data)
         return df._repr_html_()
-
-
-@dataclass
-class Code(Jsonable):
-    name: str
-    func_name: str
-    syft_link: SyftLink | None = None
-    readme_link: SyftLink | None = None
-    requirements_link: SyftLink | None = None
-    requirements: dict[str, str] | None = None
-    _func: Callable | None = None
-    _client_config: ClientConfig | None = None
-
-    def _repr_html_(self):
-        import pandas as pd
-
-        output = f"<strong>{self.name}</strong>\n"
-        table_data = {
-            "Attribute": ["Name", "Syft Link", "Readme", "Requirements"],
-            "Value": [
-                self.name,
-                "..." + str(self.syft_link)[-20:],
-                "..." + str(self.readme_link)[-20:],
-                "..." + str(self.requirements_link)[-20:],
-            ],
-        }
-
-        # Create a DataFrame from the transposed data
-        df = pd.DataFrame(table_data)
-        if self._client_config:
-            readme = self._client_config.resolve_link(self.readme_link)
-            with open(readme) as f:
-                output += "\nREADME:\n" + markdown_to_html(f.read()) + "\n"
-
-        return output + df._repr_html_()
-
-    # can also do from df where you specify the destination
-    @classmethod
-    def from_func(self, func: Callable):
-        name = func.__name__
-        code = Code(func_name=name, _func=func, name=name)
-        # code.write_files(manifest)
-        return code
-
-    def get_function_source(self, func):
-        source_code = inspect.getsource(func)
-        dedented_code = textwrap.dedent(source_code)
-        dedented_code = dedented_code.strip()
-        decorator = "@syftbox_code"
-        if dedented_code.startswith(decorator):
-            dedented_code = dedented_code[len(decorator) :]
-        return dedented_code
-
-    @property
-    def import_string(self) -> str:
-        string = "from syftbox.lib."
-        string += create_datasite_import_path(self.syft_link.datasite)  # a.at.b.com
-        string += ".code import "
-        string += self.clean_name
-        return string
-
-    def readme_template(self) -> str:
-        readme = f"""
-        # {self.name}
-
-        Code:
-
-        ## Import Syntax
-        client_config.use()
-        {self.import_string}
-
-        ## Python Usage Example
-        result = {self.func_name}()
-        """
-        return textwrap.dedent(readme)
-
-    def __call__(self, *args, **kwargs):
-        return self._func(*args, **kwargs)
-
-    @property
-    def raw_code(self) -> str:
-        if self._func:
-            return self.get_function_source(self._func)
-
-        code = ""
-        if self._client_config:
-            code_link = self._client_config.resolve_link(self.syft_link)
-            with open(code_link) as f:
-                code = f.read()
-        return code
-
-    @property
-    def code(self):
-        from IPython.display import Markdown
-
-        return Markdown(f"```python\n{self.raw_code}\n```")
-
-    def run(self, *args, resolve_private: bool = False, **kwargs):
-        # todo figure out how to override sy_path in the sub code
-        if self._client_config:
-            code = self.raw_code
-
-            # Evaluate the code and store the function in memory
-            local_vars = {}
-            exec(code, {}, local_vars)
-
-            # Get the function name
-            function_name = f"{self.clean_name}"
-            if function_name not in local_vars:
-                raise ValueError(
-                    f"Function {function_name} not found in the loader code."
-                )
-
-            # Get the function from the local_vars
-            inner_function = local_vars[function_name]
-
-            return inner_function(*args, **kwargs)
-        else:
-            raise Exception("run client_config.use()")
-
-    @classmethod
-    def extract_imports(cls, source_code):
-        imports = set()
-        tree = ast.parse(source_code)
-        for node in ast.walk(tree):
-            if isinstance(node, ast.Import):
-                for alias in node.names:
-                    imports.add(alias.name.split(".")[0])
-            elif isinstance(node, ast.ImportFrom):
-                imports.add(node.module.split(".")[0])
-        return imports
-
-    @property
-    def clean_name(self) -> str:
-        return to_safe_function_name(self.name)
-
-    def write_files(self, manifest) -> bool:
-        code_dir = Path(manifest.root_dir / "code" / self.clean_name)
-        os.makedirs(code_dir, exist_ok=True)
-
-        manifest.create_public_folder(code_dir)
-
-        # write code
-        code_path = code_dir / (self.clean_name + ".py")
-        source = self.get_function_source(self._func)
-        with open(code_path, "w") as f:
-            f.write(source)
-        self.syft_link = SyftLink.from_path(code_path)
-
-        # write readme
-        readme_link = code_dir / "README.md"
-        with open(readme_link, "w") as f:
-            f.write(self.readme_template())
-        self.readme_link = SyftLink.from_path(readme_link)
-
-        # write requirements.txt
-        imports = self.extract_imports(source)
-        requirements = {}
-
-        installed_packages = {pkg.key: pkg.version for pkg in pkg_resources.working_set}
-        requirements_txt_path = code_dir / "requirements.txt"
-        with open(requirements_txt_path, "w") as f:
-            for package in imports:
-                if package in installed_packages:
-                    requirements[package] = installed_packages[package]
-                    f.write(f"{package}=={installed_packages[package]}\n")
-                else:
-                    requirements[package] = ""
-                    f.write(f"{package}\n")
-                    print(
-                        f"Warning: {package} is not installed in the current environment."
-                    )
-        self.requirements_link = SyftLink.from_path(requirements_txt_path)
-        self.requirements = requirements
-
-    def publish(self, manifest: DatasiteManifest, overwrite: bool = False):
-        if self.name in manifest.code and not overwrite:
-            raise Exception(f"Code: {self.name} already in manifest")
-        self.write_files(manifest)
-        manifest.code[self.name] = self.to_dict()
-        manifest.save(manifest.file_path)
-        print("✅ Code Published")
-
-    @property
-    def file_path(self):
-        if self._client_config:
-            return self._client_config.resolve_link(self.syft_link)
-
-    def to_flow(
-        self,
-        client_config,
-        inputs=None,
-        output=None,
-        template="python",
-        path=None,
-        write_back_approved_path: str | None = None,
-        write_back_denied_path: str | None = None,
-    ) -> str:
-        if path is None:
-            path = Path(client_config.sync_folder) / "staging"
-            os.makedirs(path, exist_ok=True)
-        if output is None:
-            output = {}
-
-        if "name" not in output:
-            output["name"] = "result"
-        if "format" not in output:
-            output["format"] = "json"
-
-        values = list(inputs.values())
-
-        first_value = values[0]
-
-        if isinstance(first_value, list):
-            emails = set([client_config.email])
-            for dataset in first_value:
-                their_email = dataset.syft_link.datasite
-                emails.add(their_email)
-
-            if "permission" not in output:
-                perm = SyftPermission(
-                    admin=list(emails),
-                    read=list(emails),
-                    write=list(emails),
-                )
-                output["permission"] = perm
-        else:
-            their_email = first_value.syft_link.datasite
-            if "permission" not in output:
-                perm = SyftPermission.theirs_with_my_read_write(
-                    their_email=their_email, my_email=client_config.email
-                )
-                output["permission"] = perm
-
-        # create folders
-        init_flow(client_config, path, self.name, inputs, output, template)
-        # save main.py
-        main_code = create_main_py(client_config, inputs, output, self)
-
-        flow_dir = Path(os.path.abspath(f"{path}/{self.name}"))
-
-        main_code_path = flow_dir / "main.py"
-        with open(main_code_path, "w") as f:
-            f.write(main_code)
-        main_shell_path = flow_dir / "run.sh"
-        main_shell_code = make_run_sh()
-        with open(main_shell_path, "w") as f:
-            f.write(main_shell_code)
-        make_executable(main_shell_path)
-
-        if write_back_approved_path is None:
-            write_back_approved_path = "results/2_approved"
-
-        if write_back_denied_path is None:
-            write_back_denied_path = "results/3_denied"
-
-        task_manifest = TaskManifest(
-            author=client_config.email,
-            result_datasite=client_config.email,
-            execution_datasite=their_email,
-            write_back_approved_path=write_back_approved_path,
-            write_back_denied_path=write_back_denied_path,
-        )
-
-        task_manifest_path = flow_dir / "manifest.json"
-        task_manifest.save(task_manifest_path)
-
-        return str(flow_dir)
 
 
 def syftbox_code(func):
