+++ default_apps/netflix/main.py
@@ -12,7 +12,7 @@ from utils import compute_file_hash, load_cache, save_cache
 
 def publish_page(output_path):
     try:
-        from syftbox.lib import ClientConfig
+        from syftbox.lib.client_config import ClientConfig
 
         config_path = os.environ.get("SYFTBOX_CLIENT_CONFIG_PATH", None)
         client_config = ClientConfig.load(config_path)
