+++ syftbox/lib/jsonable.py
@@ -1,0 +1,63 @@
+import json
+from typing import Any
+
+from typing_extensions import Self
+
+
+def is_primitive_json_serializable(obj):
+    if isinstance(obj, (str, int, float, bool, type(None))):
+        return True
+    return False
+
+
+def pack(obj) -> Any:
+    if is_primitive_json_serializable(obj):
+        return obj
+
+    if hasattr(obj, "to_dict"):
+        return obj.to_dict()
+
+    if isinstance(obj, list):
+        return [pack(val) for val in obj]
+
+    if isinstance(obj, dict):
+        return {k: pack(v) for k, v in obj.items()}
+
+    raise Exception(f"Unable to pack type: {type(obj)} value: {obj}")
+
+
+class Jsonable:
+    def to_dict(self) -> dict:
+        output = {}
+        for k, v in self.__dict__.items():
+            if k.startswith("_"):
+                continue
+            output[k] = pack(v)
+        return output
+
+    def __iter__(self):
+        for key, val in self.to_dict().items():
+            if key.startswith("_"):
+                yield key, val
+
+    def __getitem__(self, key):
+        if key.startswith("_"):
+            return None
+        return self.to_dict()[key]
+
+    @classmethod
+    def load(cls, filepath: str) -> Self:
+        try:
+            with open(filepath) as f:
+                data = f.read()
+                d = json.loads(data)
+                return cls(**d)
+        except Exception as e:
+            raise e
+            print(f"Unable to load jsonable file: {filepath}. {e}")
+        return None
+
+    def save(self, filepath: str) -> None:
+        d = self.to_dict()
+        with open(filepath, "w") as f:
+            f.write(json.dumps(d))
