+++ syftbox/lib/link.py
@@ -1,0 +1,174 @@
+import copy
+import os
+import re
+from dataclasses import dataclass
+from urllib.parse import urlparse
+
+import requests
+from typing_extensions import Self
+
+from .jsonable import Jsonable
+from .util import extract_leftmost_email, validate_email, verify_tls
+
+
+@dataclass
+class SyftLink(Jsonable):
+    @classmethod
+    def from_file(cls, path: str) -> Self:
+        if not os.path.exists(path):
+            raise Exception(f"{path} does not exist")
+        with open(path, "r") as f:
+            return cls.from_url(f.read())
+
+    def from_path(path: str) -> Self | None:
+        parts = []
+        collect = False
+        for part in str(path).split("/"):
+            # quick hack find the first email and thats the datasite
+            if collect:
+                parts.append(part)
+            elif validate_email(part):
+                collect = True
+                parts.append(part)
+
+        if len(parts):
+            sync_path = "/".join(parts)
+            return SyftLink.from_url(f"syft://{sync_path}")
+        return None
+
+    @classmethod
+    def from_url(cls, url: str | "SyftLink") -> Self:
+        if isinstance(url, SyftLink):
+            return url
+        try:
+            # urlparse doesnt handle no protocol properly
+            if "://" not in url:
+                url = "http://" + url
+            parts = urlparse(url)
+            host_or_ip_parts = parts.netloc.split(":")
+            # netloc is host:port
+            port = 80
+            if len(host_or_ip_parts) > 1:
+                port = int(host_or_ip_parts[1])
+            host_or_ip = host_or_ip_parts[0]
+            if parts.scheme == "https":
+                port = 443
+
+            return SyftLink(
+                host_or_ip=host_or_ip,
+                path=parts.path,
+                port=port,
+                protocol=parts.scheme,
+                query=getattr(parts, "query", ""),
+            )
+        except Exception as e:
+            raise e
+
+    def to_file(self, path: str) -> bool:
+        with open(path, "w") as f:
+            f.write(str(self))
+
+    def __init__(
+        self,
+        protocol: str = "http",
+        host_or_ip: str = "localhost",
+        port: int | None = 5001,
+        path: str = "",
+        query: str = "",
+    ) -> None:
+        # in case a preferred port is listed but its not clear if an alternative
+        # port was included in the supplied host_or_ip:port combo passed in earlier
+        match_port = re.search(":[0-9]{1,5}", host_or_ip)
+        if match_port:
+            sub_server_url: SyftLink = SyftLink.from_url(host_or_ip)
+            host_or_ip = str(sub_server_url.host_or_ip)  # type: ignore
+            port = int(sub_server_url.port)  # type: ignore
+            protocol = str(sub_server_url.protocol)  # type: ignore
+            path = str(sub_server_url.path)  # type: ignore
+
+        prtcl_pattrn = "://"
+        if prtcl_pattrn in host_or_ip:
+            protocol = host_or_ip[: host_or_ip.find(prtcl_pattrn)]
+            start_index = host_or_ip.find(prtcl_pattrn) + len(prtcl_pattrn)
+            host_or_ip = host_or_ip[start_index:]
+
+        self.host_or_ip = host_or_ip
+        self.path: str = path
+        self.port = port
+        self.protocol = protocol
+        self.query = query
+
+    def with_path(self, path: str) -> Self:
+        dupe = copy.copy(self)
+        dupe.path = path
+        return dupe
+
+    @property
+    def query_string(self) -> str:
+        query_string = ""
+        if len(self.query) > 0:
+            query_string = f"?{self.query}"
+        return query_string
+
+    @property
+    def url(self) -> str:
+        return f"{self.base_url}{self.path}{self.query_string}"
+
+    @property
+    def url_no_port(self) -> str:
+        return f"{self.base_url_no_port}{self.path}{self.query_string}"
+
+    @property
+    def base_url(self) -> str:
+        return f"{self.protocol}://{self.host_or_ip}:{self.port}"
+
+    @property
+    def base_url_no_port(self) -> str:
+        return f"{self.protocol}://{self.host_or_ip}"
+
+    @property
+    def url_no_protocol(self) -> str:
+        return f"{self.host_or_ip}:{self.port}{self.path}"
+
+    @property
+    def url_path(self) -> str:
+        return f"{self.path}{self.query_string}"
+
+    def to_tls(self) -> Self:
+        if self.protocol == "https":
+            return self
+
+        # TODO: only ignore ssl in dev mode
+        r = requests.get(  # nosec
+            self.base_url, verify=verify_tls()
+        )  # ignore ssl cert if its fake
+        new_base_url = r.url
+        if new_base_url.endswith("/"):
+            new_base_url = new_base_url[0:-1]
+        return self.__class__.from_url(
+            url=f"{new_base_url}{self.path}{self.query_string}"
+        )
+
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} {self.url}>"
+
+    def __str__(self) -> str:
+        return self.url
+
+    def __hash__(self) -> int:
+        return hash(self.__str__())
+
+    def __copy__(self) -> Self:
+        return self.__class__.from_url(self.url)
+
+    def set_port(self, port: int) -> Self:
+        self.port = port
+        return self
+
+    @property
+    def sync_path(self) -> str:
+        return self.host_or_ip + self.path
+
+    @property
+    def datasite(self) -> str:
+        return extract_leftmost_email(str(self))
